multiple_physics_pretraining_paddle/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

adan-pypaddlepaddle-gpu==0.1.0
dadaptation==3.1
einops==0.7
h5py==3.9.0
numpy==1.25.2
timm==0.9.5
paddlepaddle-gpu==2.1
paddlepaddle-gpuinfo==1.8.0
wandb==0.15.8
tqdm==4.66.0
zarr==2.16.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
